packages/mindgraph/python/requirements.txt
graphiti-core==0.28.2
neo4j==5.26.0
python-dotenv==1.0.1
openai==1.76.0
numpy==2.2.0
